+++ asgardpy/data/dataset_3d.py
@@ -25,21 +25,23 @@ from gammapy.modeling.models import (
     TemplateSpatialModel,
 )
 
-from asgardpy.base import (
-    AnalysisStepBase,
-    BackgroundConfig,
-    BaseConfig,
+from asgardpy.analysis.step_base import AnalysisStepBase
+from asgardpy.base.base import BaseConfig
+from asgardpy.base.geom import (
     GeomConfig,
     MapAxesConfig,
+    SkyPositionConfig,
+    create_counts_map,
+    generate_geom,
+    get_energy_axis,
+    get_source_position,
+)
+from asgardpy.base.reduction import (
+    BackgroundConfig,
     MapSelectionEnum,
     ObservationsConfig,
     ReductionTypeEnum,
     SafeMaskConfig,
-    SkyPositionConfig,
-    get_energy_axis,
-)
-from asgardpy.base.geom import create_counts_map, generate_geom, get_source_position
-from asgardpy.base.reduction import (
     generate_dl4_dataset,
     get_bkg_maker,
     get_dataset_maker,