+++ asgardpy/io/__init__.py
@@ -1,6 +1,9 @@
 """
 Input/Output Module
-"""
-from asgardpy.io.io import DL3Files, InputConfig, InputFilePatterns
 
-__all__ = ["DL3Files", "InputConfig", "InputFilePatterns"]
+# order matters to prevent circular imports
+isort:skip_file
+"""
+from asgardpy.io.io import InputConfig, InputFilePatterns, DL3Files
+
+__all__ = ["InputConfig", "InputFilePatterns", "DL3Files"]