+++ asgardpy/stats/__init__.py
@@ -1,5 +1,8 @@
 """
 Statistics Module
+
+# order matters to prevent circular imports
+isort:skip_file
 """
 from asgardpy.stats.stats import (
     check_model_preference_aic,